rename module bapsflib._hdf.maps.hdfmap -> bapsflib._hdf.maps.core
