+++ bapsflib/_hdf/maps/core.py
