+++ docs/api_static/bapsflib._hdf.maps.core.rst
@@ -1,0 +1,9 @@
+bapsflib\.\_hdf\.maps\.core
+===========================
+
+.. py:currentmodule:: bapsflib._hdf.maps.core
+
+.. automodapi:: bapsflib._hdf.maps.core
+    :no-heading:
+    :include-all-objects:
+    :headings: "-^"
