--- docs/api_static/bapsflib._hdf.maps.hdfmap.rst
@@ -1,9 +1,0 @@
-bapsflib\.\_hdf\.maps\.hdfmap
-=============================
-
-.. py:currentmodule:: bapsflib._hdf.maps.hdfmap
-
-.. automodapi:: bapsflib._hdf.maps.hdfmap
-    :no-heading:
-    :include-all-objects:
-    :headings: "-^"
