+++ docs/api_static/bapsflib._hdf.maps.rst
@@ -13,7 +13,7 @@ Sub-Packages & Modules
 
     controls
     digitizers
-    hdfmap
+    core
     msi
 
 .. automodapi:: bapsflib._hdf.maps
